Refactor ExportDialog pages into separate components
