+++ apps/src/code-studio/components/ExportDialog/styles.js
@@ -1,0 +1,27 @@
+import color from '../../../util/color';
+
+export default {
+  button: {
+    borderWidth: 1,
+    borderColor: color.border_gray,
+    fontSize: 'larger',
+    outline: 'none',
+    padding: 10,
+    marginTop: 0,
+    marginBottom: 0,
+    marginLeft: 0,
+    marginRight: 8,
+    verticalAlign: 'top'
+  },
+  section: {
+    marginTop: 10
+  },
+  text: {
+    fontSize: 'inherit',
+    lineHeight: 'inherit',
+    color: 'inherit'
+  },
+  title: {
+    fontSize: 16
+  }
+};
